Virtual-machine3-web/requirements.txt
psycopg2-binary>=2.9.1
pandas>=1.3.3
flask
